# examples/petstore/openapi.yaml
openapi: 3.0.3
info:
  title: Petstore
  description: A sample API that demonstrates pyoas code generation from an OpenAPI spec.
  version: 1.0.0

tags:
  - name: Pets
    description: Operations about pets
  - name: Store
    description: Access to petstore orders
  - name: Users
    description: Operations about users

components:
  schemas:
    Pet:
      description: A pet in the store
      type: object
      required: [id, name, status]
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        name:
          type: string
          description: The pet's name
        status:
          type: string
          enum: [available, pending, sold]
          default: available
        tag:
          type: string
          nullable: true

    CreatePetRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        tag:
          type: string
          nullable: true

    UpdatePetRequest:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [available, pending, sold]
        tag:
          type: string
          nullable: true

    PetList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Pet'
        total:
          type: integer

    Order:
      type: object
      required: [id, pet_id, quantity, status]
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        pet_id:
          type: integer
          format: int64
        quantity:
          type: integer
          minimum: 1
          maximum: 100
        status:
          type: string
          enum: [placed, approved, delivered]
        ship_date:
          type: string
          format: date-time
          nullable: true

    CreateOrderRequest:
      type: object
      required: [pet_id, quantity]
      properties:
        pet_id:
          type: integer
          format: int64
        quantity:
          type: integer
          minimum: 1
          maximum: 100

    User:
      type: object
      required: [id, username, email]
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        username:
          type: string
        email:
          type: string
        first_name:
          type: string
          nullable: true
        last_name:
          type: string
          nullable: true

    CreateUserRequest:
      type: object
      required: [username, email]
      properties:
        username:
          type: string
        email:
          type: string
        first_name:
          type: string
          nullable: true
        last_name:
          type: string
          nullable: true

paths:
  /pets:
    get:
      operationId: listPets
      summary: List all pets
      tags: [Pets]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: status
          in: query
          schema:
            type: string
            enum: [available, pending, sold]
      responses:
        '200':
          description: A paginated list of pets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetList'

    post:
      operationId: createPet
      summary: Create a pet
      tags: [Pets]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePetRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'

  /pets/{petId}:
    get:
      operationId: getPet
      summary: Get a pet by ID
      tags: [Pets]
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: The pet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        '404':
          description: Pet not found

    patch:
      operationId: updatePet
      summary: Update a pet
      tags: [Pets]
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePetRequest'
      responses:
        '200':
          description: Updated pet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        '404':
          description: Pet not found

    delete:
      operationId: deletePet
      summary: Delete a pet
      tags: [Pets]
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Deleted

  /store/inventory:
    get:
      operationId: getInventory
      summary: Get store inventory
      tags: [Store]
      responses:
        '200':
          description: Inventory counts by status
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: integer

  /store/orders:
    post:
      operationId: createOrder
      summary: Place an order for a pet
      tags: [Store]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order placed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

  /store/orders/{orderId}:
    get:
      operationId: getOrder
      summary: Get order by ID
      tags: [Store]
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: The order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found

  /users:
    post:
      operationId: createUser
      summary: Create a user account
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  /users/{username}:
    get:
      operationId: getUser
      summary: Get user by username
      tags: [Users]
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found