# examples/petstore/src/generated/models/pets.py
# Generated by pyoas — do not edit manually. spec: 91555b2d
from __future__ import annotations

from typing import Literal

from pydantic import BaseModel, ConfigDict, Field


class PetList(BaseModel):
    model_config = ConfigDict(
        extra="ignore",
        frozen=False,
        populate_by_name=True,
    )

    items: list[Pet] = Field(default=...)
    total: int = Field(default=...)


class PetRead(BaseModel):
    """A pet in the store"""

    model_config = ConfigDict(
        extra="ignore",
        frozen=False,
        populate_by_name=True,
    )

    id: int = Field(default=...)
    name: str = Field(default=..., description="The pet's name")
    status: Literal["available", "pending", "sold"] = Field(default="available")
    tag: str | None = Field(default=None)


class PetWrite(BaseModel):
    """A pet in the store"""

    model_config = ConfigDict(
        extra="forbid",
        frozen=False,
        populate_by_name=True,
    )

    name: str = Field(default=..., description="The pet's name")
    status: Literal["available", "pending", "sold"] = Field(default="available")
    tag: str | None = Field(default=None)


class CreatePetRequest(BaseModel):
    model_config = ConfigDict(
        extra="forbid",
        frozen=False,
        populate_by_name=True,
    )

    name: str = Field(default=...)
    tag: str | None = Field(default=None)


class UpdatePetRequest(BaseModel):
    model_config = ConfigDict(
        extra="forbid",
        frozen=False,
        populate_by_name=True,
    )

    name: str | None = Field(default=None)
    status: Literal["available", "pending", "sold"] | None = Field(default=None)
    tag: str | None = Field(default=None)


Pet = PetRead