# examples/petstore/src/generated/models/store.py
# Generated by pyoas — do not edit manually. spec: 91555b2d
from __future__ import annotations

import datetime
from typing import Annotated, Literal

from pydantic import BaseModel, ConfigDict, Field


class CreateOrderRequest(BaseModel):
    model_config = ConfigDict(
        extra="forbid",
        frozen=False,
        populate_by_name=True,
    )

    pet_id: int = Field(default=...)
    quantity: Annotated[int, Field(ge=1, le=100, default=...)]


class OrderRead(BaseModel):
    model_config = ConfigDict(
        extra="ignore",
        frozen=False,
        populate_by_name=True,
    )

    id: int = Field(default=...)
    pet_id: int = Field(default=...)
    quantity: Annotated[int, Field(ge=1, le=100, default=...)]
    status: Literal["placed", "approved", "delivered"] = Field(default=...)
    ship_date: datetime.datetime | None = Field(default=None)


class OrderWrite(BaseModel):
    model_config = ConfigDict(
        extra="forbid",
        frozen=False,
        populate_by_name=True,
    )

    pet_id: int = Field(default=...)
    quantity: Annotated[int, Field(ge=1, le=100, default=...)]
    status: Literal["placed", "approved", "delivered"] = Field(default=...)
    ship_date: datetime.datetime | None = Field(default=None)


Order = OrderRead