# examples/petstore/src/generated/models/users.py
# Generated by pyoas — do not edit manually. spec: 91555b2d
from __future__ import annotations

from pydantic import BaseModel, ConfigDict, Field


class CreateUserRequest(BaseModel):
    model_config = ConfigDict(
        extra="forbid",
        frozen=False,
        populate_by_name=True,
    )

    username: str = Field(default=...)
    email: str = Field(default=...)
    first_name: str | None = Field(default=None)
    last_name: str | None = Field(default=None)


class UserRead(BaseModel):
    model_config = ConfigDict(
        extra="ignore",
        frozen=False,
        populate_by_name=True,
    )

    id: int = Field(default=...)
    username: str = Field(default=...)
    email: str = Field(default=...)
    first_name: str | None = Field(default=None)
    last_name: str | None = Field(default=None)


class UserWrite(BaseModel):
    model_config = ConfigDict(
        extra="forbid",
        frozen=False,
        populate_by_name=True,
    )

    username: str = Field(default=...)
    email: str = Field(default=...)
    first_name: str | None = Field(default=None)
    last_name: str | None = Field(default=None)


User = UserRead