# examples/petstore/src/generated/routers/pets.py
# Generated by pyoas — do not edit manually. spec: 91555b2d
from __future__ import annotations

from typing import Annotated, Literal

from fastapi import APIRouter, Body, Depends, Path, Query
from services.pets import PetsService, get_pets_service

from generated.models.pets import CreatePetRequest, Pet, PetList, UpdatePetRequest

router = APIRouter(tags=["Pets"])


@router.get("/pets", response_model=PetList, summary="List all pets")
async def list_pets(
    *,
    limit: Annotated[int | None, Query(ge=1, le=100)] = None,
    status: Literal["available", "pending", "sold"] | None = None,
    service: PetsService = Depends(get_pets_service),
) -> PetList:
    return await service.list_pets(limit=limit, status=status)


@router.post("/pets", response_model=Pet, status_code=201, summary="Create a pet")
async def create_pet(
    *,
    body: Annotated[CreatePetRequest, Body()],
    service: PetsService = Depends(get_pets_service),
) -> Pet:
    return await service.create_pet(body=body)


@router.get("/pets/{petId}", response_model=Pet, summary="Get a pet by ID")
async def get_pet(
    *,
    pet_id: Annotated[int, Path(alias="petId")],
    service: PetsService = Depends(get_pets_service),
) -> Pet:
    return await service.get_pet(pet_id=pet_id)


@router.patch("/pets/{petId}", response_model=Pet, summary="Update a pet")
async def update_pet(
    *,
    pet_id: Annotated[int, Path(alias="petId")],
    body: Annotated[UpdatePetRequest, Body()],
    service: PetsService = Depends(get_pets_service),
) -> Pet:
    return await service.update_pet(pet_id=pet_id, body=body)


@router.delete("/pets/{petId}", status_code=204, summary="Delete a pet")
async def delete_pet(
    *,
    pet_id: Annotated[int, Path(alias="petId")],
    service: PetsService = Depends(get_pets_service),
) -> None:
    await service.delete_pet(pet_id=pet_id)