# examples/petstore/src/generated/routers/store.py
# Generated by pyoas — do not edit manually. spec: 91555b2d
from __future__ import annotations

from typing import Annotated

from fastapi import APIRouter, Body, Depends, Path
from services.store import StoreService, get_store_service

from generated.models.store import CreateOrderRequest, Order

router = APIRouter(tags=["Store"])


@router.get(
    "/store/inventory", response_model=dict[str, int], summary="Get store inventory"
)
async def get_inventory(
    *,
    service: StoreService = Depends(get_store_service),
) -> dict[str, int]:
    return await service.get_inventory()


@router.post(
    "/store/orders",
    response_model=Order,
    status_code=201,
    summary="Place an order for a pet",
)
async def create_order(
    *,
    body: Annotated[CreateOrderRequest, Body()],
    service: StoreService = Depends(get_store_service),
) -> Order:
    return await service.create_order(body=body)


@router.get("/store/orders/{orderId}", response_model=Order, summary="Get order by ID")
async def get_order(
    *,
    order_id: Annotated[int, Path(alias="orderId")],
    service: StoreService = Depends(get_store_service),
) -> Order:
    return await service.get_order(order_id=order_id)