# examples/petstore/src/generated/routers/users.py
# Generated by pyoas — do not edit manually. spec: 91555b2d
from __future__ import annotations

from typing import Annotated

from fastapi import APIRouter, Body, Depends
from services.users import UsersService, get_users_service

from generated.models.users import CreateUserRequest, User

router = APIRouter(tags=["Users"])


@router.post(
    "/users", response_model=User, status_code=201, summary="Create a user account"
)
async def create_user(
    *,
    body: Annotated[CreateUserRequest, Body()],
    service: UsersService = Depends(get_users_service),
) -> User:
    return await service.create_user(body=body)


@router.get("/users/{username}", response_model=User, summary="Get user by username")
async def get_user(
    *,
    username: str,
    service: UsersService = Depends(get_users_service),
) -> User:
    return await service.get_user(username=username)